requirements.txt
allure-pytest==2.15.3
pytest==9.0.1
selene==2.0.0rc9
allure-python-commons~=2.15.3
selenium~=4.39.0

